Fixes #65 (change collision check to the outer circle)
As a matter of fact this is not changing the collision check at all. As mentioned in the issue, it is changing how the draw function was making use of the data members of the class, that I don't see the point in doing it like so. Whatever other stylistic fixes we need to do, we should do it on top of the data members and not on or with the data the members.