Implements both portrait and landscape mode (#89)
Had to restructure the code quite a bit for this one. The difficult part
was to be able to maintain the state (position of the particles
basically) while also drawing everything correctly. It is a simple as
having a different update cycle when orientation happens. The way to
manage this easily in code was to encapsulate as much behaviour as
possible inside the classes (Module, ModuleConsole, Button and Fader)
and outside the main ofApp class with little repetition. For that I used
a .setDimensions() method "pattern" that is both called by the
constructor of the class when starting the app as well as when
orientation changes. I do that for all the four aforementioned classes
and that is pretty much it.

Apart from that, so minor bug fixes included:

- the modules were starting with wrong values of y, namely the first one
  with y=0, the second one with y=1, and the following two ones with y=2
  and y=3, which generated some minor yet noticeably visible artifacts.

- clearer convention for defines and helper defines too;

- better type safe signatures methods: was using float when should
  clearly be int;

- removed some obsolete code: icons stuff, different drawing routines,
  etc.