Fixes for iPad
It is quite a problem not being able to detect if running for iPad or
iPhone in the preprocessing stage.. As of right now, because the modules
are a static variable, even for iPhone we need to initialize with 4
modules, en though we just need one. Should try and get a fix for that.