Fixes #115 (wrong boundary for buttons)
We were keeping logic in the draw function, instead of defining it in
the constructors. That way, the draw() method doesn't need to know about
calculating the sizes, which makes a lot of sense. It just needs to know
how to draw a rect and a title and that's it.

Minor refactor in other pieces of code for clarity.